Admin files organized and done Certain changes which we Discuss!!
